[2716] Fixed: Unix/Linux (now really) after last quest patch. Many cleanups and code fixes to make GCC happy to not output warnings. Fixed: typo in item enchanting code.

+++ src/game/MapManager.cpp
@@ -57,7 +57,7 @@ MapManager::~MapManager()
     for(MapMapType::iterator iter=i_maps.begin(); iter != i_maps.end(); ++iter)
         delete iter->second;
 
-    for(int i = 0; i < m_Transports.size(); i++)
+    for(size_t i = 0; i < m_Transports.size(); i++)
         delete m_Transports[i];
 
     sDatabase.PExecute("TRUNCATE table `creature_grid`");
